Fix: Unit tests (file dependencies), refactor folder structure
