+++ server/serverAPI/middleware/authentication/IAuthenticator.ts
